+++ Modules/Intent.Modules.Application.DomainInteractions/InteractionStrategies/UpdateEntityInteractionStrategy.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Intent.Exceptions;
+﻿using Intent.Exceptions;
 using Intent.Metadata.Models;
 using Intent.Modelers.Domain.Api;
 using Intent.Modelers.Services.DomainInteractions.Api;
@@ -11,7 +8,11 @@ using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Interactions;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
+using Intent.Modules.Common.UnitOfWork.Settings;
 using Intent.Modules.Constants;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Intent.Modules.Application.DomainInteractions.InteractionStrategies
 {
@@ -35,13 +36,13 @@ namespace Intent.Modules.Application.DomainInteractions.InteractionStrategies
                 ? queryContext.GetDtoProjectionReturnType()
                 : null;
 
-            method.AddStatements(method.GetQueryStatements(
+            method.AddStatements(ExecutionPhases.Retrieval, method.GetQueryStatements(
                 dataAccessProvider: dataAccess,
                 interaction: interaction,
                 foundEntity: foundEntity,
                 projectedType: projectedType));
 
-            method.AddStatement(string.Empty);
+            method.AddStatement(ExecutionPhases.Retrieval, string.Empty);
 
             var updateAction = interaction.AsUpdateEntityActionTargetEndModel();
             var csharpMapping = method.GetMappingManager();
@@ -92,16 +93,16 @@ namespace Intent.Modules.Application.DomainInteractions.InteractionStrategies
                     statements.Add(entityDetails.DataAccessProvider.Update(entityDetails.VariableName)
                         .SeparatedFromPrevious());
                 }
-                else // Unit Of Work Module Auto Save Changes disabled?
-                if (bool.TryParse(method.Class.File.Template.ExecutionContext.GetSettings().GetGroup("c4b7e545-eaac-42bc-8f06-2768ac8dad99").GetSetting("d6338b7c-b0f9-46bd-8dbb-3c745d5f8623").Value, out bool uowAutoSaveOff))
-                {
-                    if (!uowAutoSaveOff)
-                    {
-                        method.AddStatement(ExecutionPhases.Persistence, new CSharpStatement($"{entityDetails.DataAccessProvider.SaveChangesAsync()}"));
-                    }
-                }
 
-                method.AddStatements(statements);
+                method.AddStatements(ExecutionPhases.BusinessLogic, statements);
+
+                var automaticallyPersistUnitOfWork = method.Class.File.Template.ExecutionContext.GetSettings()
+                    .GetUnitOfWorkSettings()?.AutomaticallyPersistUnitOfWork() ?? true;
+                var saveAlreadyCalled = method.FindStatement(x => x.ToString().Trim().Contains(entityDetails.DataAccessProvider.SaveChangesAsync().ToString().Trim())) != null;
+                if (!saveAlreadyCalled && !automaticallyPersistUnitOfWork)
+                {
+                    method.AddStatement(ExecutionPhases.Persistence, new CSharpStatement($"{entityDetails.DataAccessProvider.SaveChangesAsync()}"));
+                }
             }
             catch (Exception ex)
             {