+++ Modules/Intent.Modules.Application.DomainInteractions/release-notes.md
@@ -1,3 +1,8 @@
+### Version 1.1.6
+
+- Improvement: Improved error message when `ProjectTo` is selected as the query strategy with an unsupporting provider (mapperly)
+- Fixed: Unnecesary dependency on `Intent.Common.UnitOfWork` module causing errors when the module isn't installed.
+
 ### Version 1.1.5
 
 - Improvement: Support for Mapperly module